chahua: 重构 renderer.js（六）—— 抽出 sidebar 模块
侧栏茶客列表 + 头像 + 打分小数字 + 「我」那一行抽到 sidebar.js。模块拥有打分明细
（scoresByName）/ 打分 span 引用（scoreSpansByName）/ 用户显示名 + 头像；茶客名单
（guests）仍归 renderer（被 mention / assign / 任务卡 owner 等一众模块读），sidebar
经 getGuests 取数。

renderer 经 makeAvatar / makeUserAvatar（喂 chat_stream）/ setUser / renderGuests /
applyScores / getUserDisplayName 交互。chat_view.js 的四个导入随 makeAvatar* /
renderGuestRowNormal 一并移入 sidebar，renderer 不再 import chat_view。

renderer.js 1110 → 986 行（自原始 1671 累计 −685，−41%）。行为零变化，
node --check 通过。
